fix: don't orphan resources that fail during creation
Fixes networks, clusters, and serverless_clusters to store their id
in the Terraform state as soon as we have it. This way, if creation
fails for any reason, Terraform knows that we have a tainted resource
and it can be deleted, manually untainted, etc., and Terraform won't
try to recreate it (which usually fails since the name is taken).